+++ src/lib.rs
@@ -131,7 +131,7 @@
 //!     Ok(())
 //! }
 //! ```
-#![doc(html_root_url = "https://docs.rs/modio/0.4.0")]
+#![doc(html_root_url = "https://docs.rs/modio/0.4.1")]
 #![deny(rust_2018_idioms)]
 
 #[macro_use]